implment input validation
